+++ wemake_python_styleguide/logics/imports.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import ast
+from typing import List
 
 from wemake_python_styleguide.types import AnyImport
 
@@ -14,3 +15,9 @@ def get_error_text(node: AnyImport) -> str:
     if isinstance(node, ast.Import):
         return node.names[0].name
     return '.'
+
+
+def get_import_parts(node: AnyImport) -> List[str]:
+    """Returns list of import modules."""
+    module_path = getattr(node, 'module', '') or ''
+    return module_path.split('.')