+++ wemake_python_styleguide/logics/variables/access.py
@@ -1,0 +1,47 @@
+# -*- coding: utf-8 -*-
+
+from typing import Optional
+
+
+def is_private_variable(name: Optional[str]) -> bool:
+    """
+    Checks if variable has private name pattern.
+
+    >>> is_private_variable(None)
+    False
+
+    >>> is_private_variable('regular')
+    False
+
+    >>> is_private_variable('__private')
+    True
+
+    >>> is_private_variable('_protected')
+    False
+
+    >>> is_private_variable('__magic__')
+    False
+
+    """
+    return (
+        name is not None and name.startswith('__') and not name.endswith('__')
+    )
+
+
+def is_protected_variable(name: str) -> bool:
+    """
+    Checks if variable has protected name pattern.
+
+    >>> is_protected_variable('_protected')
+    True
+
+    >>> is_protected_variable('__private')
+    False
+
+    >>> is_protected_variable('__magic__')
+    False
+
+    >>> is_protected_variable('common_variable')
+    False
+    """
+    return name.startswith('_') and not name.startswith('__')