+++ wemake_python_styleguide/logics/variables/naming.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 from typing import Iterable, Optional
 
 from wemake_python_styleguide import constants