+++ wemake_python_styleguide/visitors/ast/imports.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 
 import ast
+from itertools import chain
 from typing import Callable
 
 from wemake_python_styleguide.constants import FUTURE_IMPORTS_WHITELIST
-from wemake_python_styleguide.logics.imports import get_error_text
+from wemake_python_styleguide.logics import imports
+from wemake_python_styleguide.logics.variables import access
 from wemake_python_styleguide.types import AnyImport, final
 from wemake_python_styleguide.violations.base import BaseViolation
 from wemake_python_styleguide.violations.best_practices import (
@@ -15,7 +17,10 @@ from wemake_python_styleguide.violations.consistency import (
     DottedRawImportViolation,
     LocalFolderImportViolation,
 )
-from wemake_python_styleguide.violations.naming import SameAliasImportViolation
+from wemake_python_styleguide.violations.naming import (
+    ProtectedModuleViolation,
+    SameAliasImportViolation,
+)
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 
 ErrorCallback = Callable[[BaseViolation], None]
@@ -29,13 +34,13 @@ class _ImportsChecker(object):
         self.error_callback = error_callback
 
     def check_nested_import(self, node: AnyImport) -> None:
-        text = get_error_text(node)
+        text = imports.get_error_text(node)
         parent = getattr(node, 'parent', None)
         if parent is not None and not isinstance(parent, ast.Module):
             self.error_callback(NestedImportViolation(node, text=text))
 
     def check_local_import(self, node: ast.ImportFrom) -> None:
-        text = get_error_text(node)
+        text = imports.get_error_text(node)
         if node.level != 0:
             self.error_callback(
                 LocalFolderImportViolation(node, text=text),
@@ -63,6 +68,15 @@ class _ImportsChecker(object):
                     SameAliasImportViolation(node, text=alias.name),
                 )
 
+    def check_protected_import(self, node: AnyImport) -> None:
+        text = imports.get_error_text(node)
+        import_names = [alias.name for alias in node.names]
+        for name in chain(imports.get_import_parts(node), import_names):
+            if access.is_protected_variable(name):
+                self.error_callback(
+                    ProtectedModuleViolation(node, text=text),
+                )
+
 
 @final
 class WrongImportVisitor(BaseNodeVisitor):
@@ -86,6 +100,7 @@ class WrongImportVisitor(BaseNodeVisitor):
         self._checker.check_nested_import(node)
         self._checker.check_dotted_raw_import(node)
         self._checker.check_alias(node)
+        self._checker.check_protected_import(node)
         self.generic_visit(node)
 
     def visit_ImportFrom(self, node: ast.ImportFrom) -> None:
@@ -103,4 +118,5 @@ class WrongImportVisitor(BaseNodeVisitor):
         self._checker.check_nested_import(node)
         self._checker.check_future_import(node)
         self._checker.check_alias(node)
+        self._checker.check_protected_import(node)
         self.generic_visit(node)